--Change Log:- Font class will now only deal with calling the FontFactory to load the font. if the font already created and loaded, FontFactory will return that font instead of creating a new one. If that font does not exist FontFactory will create one and return it. -Fixed a bug where the text position would be offset by few extra pixels depending on the text size. Therefor the position was never correct. Before when you set the text to 0, 0 the text position would actually be something like 0, 20. Now when you set the text position to 0, 0, the text would be at 0, 0. -StyleSheet now can load default font and text and text alignment for GUI. -Text is now rendered by creating a Text object from the Text class and passing it to the spriteRenderer to be rendered. -Now in order to render Text you need a Font and a Text objects. -Added Label class that deals with rendering the GUI Text. ***This is work in progress***

+++ Blue-Flame-Engine/Core/BF/Platform/API/DirectX/DXContext.cpp
@@ -239,8 +239,8 @@ namespace BF
 
 					viewPort.TopLeftX = 0;
 					viewPort.TopLeftY = 0;
-					viewPort.Width = Engine::GetWindow().GetClientWidth();
-					viewPort.Height = Engine::GetWindow().GetClientHeight();
+					viewPort.Width = (float)Engine::GetWindow().GetClientWidth();
+					viewPort.Height = (float)Engine::GetWindow().GetClientHeight();
 					viewPort.MinDepth = 0.0f;
 					viewPort.MaxDepth = 1.0f;
 					context->RSSetViewports(1, &viewPort);
